requirements.txt
aiogram==3.22.0
aiosqlite==0.21.0
APScheduler==3.11.0
python-dotenv==1.1.1
tzdata==2025.2
